refactor stocks note
