+++ src/main/java/uk/gov/companieshouse/api/accounts/model/entity/notes/stocks/StocksEntity.java
@@ -1,10 +1,10 @@
 package uk.gov.companieshouse.api.accounts.model.entity.notes.stocks;
 
 import org.springframework.data.mongodb.core.mapping.Document;
-import uk.gov.companieshouse.api.accounts.model.entity.BaseEntity;
+import uk.gov.companieshouse.api.accounts.model.entity.NoteEntity;
 
 @Document(collection = "notes")
-public class StocksEntity extends BaseEntity {
+public class StocksEntity extends NoteEntity {
 
     private StocksDataEntity data;
