+++ src/main/java/uk/gov/companieshouse/api/accounts/model/rest/smallfull/notes/stocks/Stocks.java
@@ -1,13 +1,15 @@
-package uk.gov.companieshouse.api.accounts.model.rest.notes.stocks;
+package uk.gov.companieshouse.api.accounts.model.rest.smallfull.notes.stocks;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonProperty;
-import uk.gov.companieshouse.api.accounts.model.rest.RestObject;
+import uk.gov.companieshouse.api.accounts.model.rest.Note;
+import uk.gov.companieshouse.api.accounts.model.rest.notes.stocks.CurrentPeriod;
+import uk.gov.companieshouse.api.accounts.model.rest.notes.stocks.PreviousPeriod;
 
 import javax.validation.Valid;
 
 @JsonInclude(JsonInclude.Include.NON_NULL)
-public class Stocks extends RestObject {
+public class Stocks extends Note {
 
     @Valid
     @JsonProperty("current_period")
