+++ src/main/java/uk/gov/companieshouse/api/accounts/transformer/StocksTransformer.java
@@ -2,16 +2,17 @@ package uk.gov.companieshouse.api.accounts.transformer;
 
 import org.springframework.beans.BeanUtils;
 import org.springframework.stereotype.Component;
+import uk.gov.companieshouse.api.accounts.enumeration.AccountingNoteType;
 import uk.gov.companieshouse.api.accounts.model.entity.notes.stocks.CurrentPeriodEntity;
 import uk.gov.companieshouse.api.accounts.model.entity.notes.stocks.PreviousPeriodEntity;
 import uk.gov.companieshouse.api.accounts.model.entity.notes.stocks.StocksDataEntity;
 import uk.gov.companieshouse.api.accounts.model.entity.notes.stocks.StocksEntity;
 import uk.gov.companieshouse.api.accounts.model.rest.notes.stocks.CurrentPeriod;
 import uk.gov.companieshouse.api.accounts.model.rest.notes.stocks.PreviousPeriod;
-import uk.gov.companieshouse.api.accounts.model.rest.notes.stocks.Stocks;
+import uk.gov.companieshouse.api.accounts.model.rest.smallfull.notes.stocks.Stocks;
 
 @Component
-public class StocksTransformer implements GenericTransformer<Stocks, StocksEntity> {
+public class StocksTransformer implements NoteTransformer<Stocks, StocksEntity> {
 
     @Override
     public StocksEntity transform(Stocks rest) {
@@ -65,5 +66,10 @@ public class StocksTransformer implements GenericTransformer<Stocks, StocksEntit
         }
 
         return stocks;
+    }
+
+    @Override
+    public AccountingNoteType getAccountingNoteType() {
+        return AccountingNoteType.SMALL_FULL_STOCKS;
     }
 }
