+++ src/main/resources/notes.yml
@@ -2,3 +2,4 @@ notes:
   small-full:
     - off-balance-sheet-arrangements
     - fixed-assets-investments
+    - stocks
